+++ src/Timestepping/simulation_stats.jl
@@ -8,6 +8,10 @@ simulation.
 
 - `reconnection_count::Int`: **total number of reconnections**;
 
+- `reconnection_length_loss::T`: accumulated **decrease of filament length due to reconnections**.
+  This is estimated from each reconnection event as the difference between the local vortex
+  lengths before and after the reconnection, using the straight segment approximation.
+
 - `filaments_removed_count::Int`: **total number of removed filaments** (this generally happens
   when the number of discretisation points becomes too small for the spatial discretisation
   to work);
@@ -19,15 +23,17 @@ simulation.
 """
 mutable struct SimulationStats{T <: AbstractFloat}
     reconnection_count       :: Int
+    reconnection_length_loss :: T
     filaments_removed_count  :: Int
     filaments_removed_length :: T
 end
 
-SimulationStats(::Type{T}) where {T} = SimulationStats(0, 0, zero(T))
+SimulationStats(::Type{T}) where {T} = SimulationStats(0, zero(T), 0, zero(T))
 
 function Base.show(io::IO, stats::SimulationStats)
     print(io, "SimulationStats:")
     print(io, "\n - reconnection_count        = ", stats.reconnection_count)
+    print(io, "\n - reconnection_length_loss  = ", stats.reconnection_length_loss)
     print(io, "\n - filaments_removed_count   = ", stats.filaments_removed_count)
     print(io, "\n - filaments_removed_length  = ", stats.filaments_removed_length)
 end